+++ components/collector/tests/source_collectors/jira/test_average_issue_lead_time.py
@@ -15,7 +15,7 @@ class JiraAverageIssueLeadTimeTest(JiraTestCase):
         status_issue = self.issue(status={"statusCategory": {"key": "done"}}, updated=updated_value)
         issues_json = {"total": 1, "issues": [status_issue]}
         response = await self.get_response(issues_json)
-        self.assert_measurement(response, value="3", entities=[self.entity(updated=updated_value, lead_time=3)])
+        self.assert_measurement(response, value="3", entities=[self.entity(updated=updated_value, lead_time="3")])
 
     async def test_exclude_status_category(self):
         """Test that issues are excluded if statusCategory is not done."""
@@ -29,9 +29,7 @@ class JiraAverageIssueLeadTimeTest(JiraTestCase):
     async def test_exclude_updated_field(self):
         """Test that issues are excluded if there is no field updated."""
         self.set_source_parameter("lookback_days", "424242")
-        status_issue = self.issue(
-            status={"statusCategory": {"key": "done"}},
-        )
+        status_issue = self.issue(status={"statusCategory": {"key": "done"}})
         issues_json = {"total": 0, "issues": [status_issue]}
         response = await self.get_response(issues_json)
         self.assert_measurement(response, value="0", entities=[])